returning float

+++ pipeline/databases/21-div.sql
@@ -1,7 +1,7 @@
 -- function SafeDiv created to divide the 1 and 2 number or return 0
 DELIMITER $$
 
-CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS DECIMAL(3, 1)
+CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS FLOAT
 BEGIN
     IF b = 0 THEN
         RETURN 0;
